requirements.txt
alembic==1.13.2
blinker==1.8.2
cffi==1.16.0
click==8.1.7
cryptography==42.0.4
Flask==2.0.1
Flask-JWT-Extended==4.4.4
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
pycparser==2.22
pycryptodome==3.20.0
PyJWT==2.8.0
SQLAlchemy==1.4.23
typing_extensions==4.12.2
Werkzeug==2.0.0
